Adjusted biomes colors

+++ src/konradn24/tml/collisions/SweepResult.java
@@ -4,15 +4,15 @@ public class SweepResult {
 
 	public boolean collided;   // czy doszło do kontaktu w [0,1]
     public boolean overlap;    // czy było nakładanie na starcie (t=0)
-    public float t;            // czas wejścia (0..1)
-    public float nx, ny;       // normal kolizji
+    public double t;            // czas wejścia (0..1)
+    public double nx, ny;       // normal kolizji
 
     public SweepResult() {
         collided = false;
         overlap = false;
-        t = 1f;
-        nx = 0f;
-        ny = 0f;
+        t = 1.0;
+        nx = 0.0;
+        ny = 0.0;
     }
 
     @Override
