+++ src/konradn24/tml/entities/statics/Opuntia1.java
@@ -5,7 +5,7 @@ import konradn24.tml.items.tools.Tool;
 
 public class Opuntia1 extends StaticEntity {
 
-	public Opuntia1(Handler handler, float x, float y) {
+	public Opuntia1(Handler handler, double x, double y) {
 		super(handler, x, y, 1, 1);
 	}
