+++ src/konradn24/tml/entities/statics/Pine1.java
@@ -5,7 +5,7 @@ import konradn24.tml.items.tools.Tool;
 
 public class Pine1 extends StaticEntity {
 
-	public Pine1(Handler handler, float x, float y) {
+	public Pine1(Handler handler, double x, double y) {
 		super(handler, x, y);
 	}
