+++ src/konradn24/tml/entities/statics/StaticEntity.java
@@ -5,11 +5,11 @@ import konradn24.tml.entities.Entity;
 
 public abstract class StaticEntity extends Entity {
 
-	public StaticEntity(Handler handler, float x, float y) {
+	public StaticEntity(Handler handler, double x, double y) {
 		super(handler, x, y);
 	}
 	
-	public StaticEntity(Handler handler, float x, float y, float width, float height) {
+	public StaticEntity(Handler handler, double x, double y, float width, float height) {
 		super(handler, x, y, width, height);
 	}
 }
