+++ src/konradn24/tml/entities/statics/Tree1.java
@@ -10,7 +10,7 @@ import konradn24.tml.items.tools.Tool;
 
 public class Tree1 extends StaticEntity {
 
-	public Tree1(Handler handler, float x, float y) {
+	public Tree1(Handler handler, double x, double y) {
 		super(handler, x, y, 2, 3);
 	}
