+++ src/konradn24/tml/entities/statics/Tree3.java
@@ -5,7 +5,7 @@ import konradn24.tml.items.tools.Tool;
 
 public class Tree3 extends StaticEntity {
 
-	public Tree3(Handler handler, float x, float y) {
+	public Tree3(Handler handler, double x, double y) {
 		super(handler, x, y, 2, 3);
 	}
