+++ src/konradn24/tml/utils/Transform.java
@@ -1,34 +1,34 @@
 package konradn24.tml.utils;
 
-import org.joml.Vector2f;
+import org.joml.Vector2d;
 
 public class Transform {
 
-	public Vector2f position;
-	public Vector2f size;
+	public Vector2d position;
+	public Vector2d size;
 	
 	public Transform() {
-		this.position = new Vector2f();
-		this.size = new Vector2f();
+		this.position = new Vector2d();
+		this.size = new Vector2d();
 	}
 	
-	public Transform(Vector2f position) {
+	public Transform(Vector2d position) {
 		this.position = position;
-		this.size = new Vector2f();
+		this.size = new Vector2d();
 	}
 
-	public Transform(Vector2f position, Vector2f size) {
+	public Transform(Vector2d position, Vector2d size) {
 		this.position = position;
 		this.size = size;
 	}
 	
-	public Transform(float x, float y, float width, float height) {
-		this.position = new Vector2f(x, y);
-		this.size = new Vector2f(width, height);
+	public Transform(double x, double y, double width, double height) {
+		this.position = new Vector2d(x, y);
+		this.size = new Vector2d(width, height);
 	}
 	
 	public Transform copy() {
-		return new Transform(new Vector2f(this.position), new Vector2f(this.size));
+		return new Transform(new Vector2d(this.position), new Vector2d(this.size));
 	}
 	
 	public void copy(Transform to) {
@@ -41,12 +41,12 @@ public class Transform {
 		this.size.set(from.size);
 	}
 	
-	public Transform addPosition(float dx, float dy) {
+	public Transform addPosition(double dx, double dy) {
 		this.position.add(dx, dy);
 		return this;
 	}
 	
-	public Transform addPosition(Vector2f v) {
+	public Transform addPosition(Vector2d v) {
 		this.position.add(v);
 		return this;
 	}
@@ -55,12 +55,12 @@ public class Transform {
 		return Math.abs(size.x) < 1e-6f && Math.abs(size.y) < 1e-6f;
 	}
 	
-	public float left()   	{ return position.x; }
-    public float right()  	{ return position.x + size.x; }
-    public float top()    	{ return position.y; }
-    public float bottom() 	{ return position.y + size.y; }
-    public float centerX() 	{ return position.x + size.x / 2; }
-    public float centerY() 	{ return position.y + size.y / 2; }
+	public double left()   	{ return position.x; }
+    public double right()  	{ return position.x + size.x; }
+    public double top()    	{ return position.y; }
+    public double bottom() 	{ return position.y + size.y; }
+    public double centerX() 	{ return position.x + size.x / 2; }
+    public double centerY() 	{ return position.y + size.y / 2; }
 
     public boolean intersects(Transform o) {
         return !(right() <= o.left() || left() >= o.right()
